+++ packages/sdk-next/src/client/BalanceClient.ts
@@ -1,9 +1,16 @@
 import { PolkadotClient } from 'polkadot-api';
 
-import { type Observable, combineLatest, map } from 'rxjs';
+import {
+  type Observable,
+  combineLatest,
+  debounceTime,
+  distinctUntilChanged,
+  map,
+} from 'rxjs';
 
+import { Papi } from '../api';
 import { SYSTEM_ASSET_ID } from '../consts';
-import { Papi } from '../provider';
+import { AssetNotFound } from '../errors';
 import { AssetAmount } from '../types';
 
 export class BalanceClient extends Papi {
@@ -13,11 +20,9 @@ export class BalanceClient extends Papi {
 
   async getBalance(account: string, assetId: number): Promise<bigint> {
     const query = this.api.query.AssetRegistry.Assets;
-
     const asset = await query.getValue(assetId);
-    if (!asset) {
-      throw Error('Fdf');
-    }
+
+    if (!asset) throw new AssetNotFound(assetId);
 
     if (asset.asset_type.type === 'Erc20') {
       return this.getErc20Balance(account, assetId);
@@ -48,16 +53,29 @@ export class BalanceClient extends Papi {
 
   subscribeBalance(address: string): Observable<AssetAmount[]> {
     const systemOb = this.subscribeSystemBalance(address);
-    const tokensOb = this.subscribeTokenBalance(address);
+    const tokensOb = this.subscribeTokensBalance(address);
+
     return combineLatest([systemOb, tokensOb]).pipe(
+      debounceTime(250),
       map((balance) => balance.flat())
+    );
+  }
+
+  // TODO: Impl proper balance check
+  subscribeErc20Balance(address: string): Observable<string> {
+    return this.api.event.EVM.Log.watch().pipe(
+      map(({ meta, payload }) => {
+        console.log(meta);
+        console.log(payload);
+        return 'yes';
+      })
     );
   }
 
   subscribeSystemBalance(address: string): Observable<AssetAmount> {
     const query = this.api.query.System.Account;
 
-    return query.watchValue(address).pipe(
+    return query.watchValue(address, 'best').pipe(
       map((balance) => {
         const { free, frozen } = balance.data;
         return {
@@ -68,14 +86,32 @@ export class BalanceClient extends Papi {
     );
   }
 
-  subscribeTokenBalance(address: string): Observable<AssetAmount[]> {
+  subscribeTokenBalance(
+    address: string,
+    assetId: number
+  ): Observable<AssetAmount> {
     const query = this.api.query.Tokens.Accounts;
 
-    return query.watchEntries(address).pipe(
+    return query.watchValue(address, assetId, 'best').pipe(
       map((balance) => {
+        const { free, frozen } = balance;
+        return {
+          id: assetId,
+          amount: free - frozen,
+        } as AssetAmount;
+      })
+    );
+  }
+
+  subscribeTokensBalance(address: string): Observable<AssetAmount[]> {
+    const query = this.api.query.Tokens.Accounts;
+
+    return query.watchEntries(address, { at: 'best' }).pipe(
+      distinctUntilChanged((_, current) => !current.deltas),
+      map(({ deltas }) => {
         const result: AssetAmount[] = [];
 
-        balance.deltas?.deleted.forEach((u) => {
+        deltas?.deleted.forEach((u) => {
           const [_, asset] = u.args;
           result.push({
             id: asset,
@@ -83,7 +119,7 @@ export class BalanceClient extends Papi {
           });
         });
 
-        balance.deltas?.upserted.forEach((u) => {
+        deltas?.upserted.forEach((u) => {
           const [_, asset] = u.args;
           const { free, frozen } = u.value;
           result.push({
@@ -97,7 +133,10 @@ export class BalanceClient extends Papi {
     );
   }
 
-  async getTokenBalanceData(account: string, assetId: number): Promise<bigint> {
+  private async getTokenBalanceData(
+    account: string,
+    assetId: number
+  ): Promise<bigint> {
     const { free, frozen } = await this.api.apis.CurrenciesApi.account(
       assetId,
       account