+++ packages/sdk-next/src/types.ts
@@ -1,12 +1,4 @@
-export interface Humanizer {
-  toHuman(): any;
-}
-
-export interface Transaction {
-  hex: string;
-  name?: string;
-  get<T>(): T;
-}
+import { XcmV3Junctions } from '@galacticcouncil/descriptors';
 
 export type Amount = {
   amount: bigint;
@@ -32,13 +24,13 @@ export interface Asset extends AssetMetadata {
   type: AssetType;
   existentialDeposit: bigint;
   isSufficient: boolean;
-  location?: any;
+  location?: XcmV3Multilocation;
   meta?: Record<string, string>;
   isWhiteListed?: boolean;
 }
 
 export interface Bond extends Asset {
-  underlyingAssetId: string;
+  underlyingAssetId: number;
   maturity: number;
 }
 
@@ -49,3 +41,8 @@ export interface ExternalAsset extends AssetMetadata {
   internalId: number;
   isWhiteListed?: boolean;
 }
+
+export type XcmV3Multilocation = {
+  parents: number;
+  interior: XcmV3Junctions;
+};