+++ packages/sdk-next/src/utils/index.ts
@@ -1,4 +1,6 @@
-export * as bnum from './bignumber';
+export * as big from './big';
+export * as evm from './evm';
 export * as fmt from './format';
 export * as json from './json';
 export * as math from './math';
+export * as xc from './xc';