+++ packages/sdk/test/script/examples/router/getBestBuy.ts
@@ -1,5 +1,5 @@
 import { ApiPromise } from '@polkadot/api';
-import { PoolService, TradeRouter, ZERO } from '@galacticcouncil/sdk';
+import { PoolService, TradeRouter, ZERO } from '../../../../src';
 
 import { PolkadotExecutor } from '../../PjsExecutor';
 import { ApiUrl } from '../../types';
@@ -8,7 +8,7 @@ class GetBestBuyPriceExample extends PolkadotExecutor {
   async script(api: ApiPromise): Promise<any> {
     const poolService = new PoolService(api);
     const router = new TradeRouter(poolService);
-    const bestBuy = await router.getBestBuy('1', '10', 10);
+    const bestBuy = await router.getBestBuy('0', '16', '0.1');
     const transaction = bestBuy.toTx(ZERO);
     console.log('Transaction hash: ' + transaction.hex);
     return bestBuy;